Admin profile make

+++ routes/auth.php
@@ -14,9 +14,9 @@ Route::prefix('auth')->group(function () {
   Route::middleware('auth:sanctum')->group(function () {
     Route::delete('/logout', [AuthController::class, 'logout']);
     Route::get('/register/links', [AuthController::class, 'getRegisterLinks']);
-    Route::post('/register/links', [AuthController::class, 'generateRegisterLink']);
-    Route::put('/register/links/{id}', [AuthController::class, 'updateRegisterLink']);
-    Route::delete('/register/links/{id}', [AuthController::class, 'deleteRegisterLink']);
+    Route::post('/register/links', [AuthController::class, 'generateRegisterLink'])->middleware(['ability:superadmin']);
+    Route::put('/register/links/{id}', [AuthController::class, 'updateRegisterLink'])->middleware(['ability:superadmin']);
+    Route::delete('/register/links/{id}', [AuthController::class, 'deleteRegisterLink'])->middleware(['ability:superadmin']);
     Route::get('/ratings', [AuthController::class, 'getRatings']);
   });
 });
